+++ src/PVOutput.Net/DependencyInjection/PVOutputServiceExtensions.cs
@@ -15,7 +15,6 @@ namespace PVOutput.Net.DependencyInjection
         /// </summary>
         /// <param name="services">The servicecollection to add the client to.</param>
         /// <param name="optionsAction">An action to configure the provided options.</param>
-        [System.Diagnostics.CodeAnalysis.SuppressMessage("Globalization", "CA1303:Do not pass literals as localized parameters", Justification = "Exception messages are non translatable for now")]
         public static void AddPVOutputClient(this IServiceCollection services, Action<PVOutputClientOptions> optionsAction)
         {
             if (optionsAction == null)