+++ src/PVOutput.Net/Objects/Modules/Implementations/BatchOutputPost.cs
@@ -1,7 +1,0 @@
-﻿namespace PVOutput.Net.Objects.Modules.Implementations
-{
-    internal sealed class BatchOutputPost : BaseOutputPost, IBatchOutputPost
-    {
-        public int? EnergyUsed { get; set; }
-    }
-}